+++ Tests/Tests/ViewSpec.swift
@@ -242,6 +242,13 @@ class ViewSpec: QuickSpec {
                     expect(cell).to(beAKindOf(CarouselCell.self))
                 }
 
+                it("it will display Carousel view and Carousel Page Control if custom Json has valid modify modifyModal size specs") {
+                    presenter.isValidSize = true
+                    view.setup(viewModel: TestHelpers.generateFullViewCarouselModel(carouselData: carouselData))
+                    expect(view.carouselView.isHidden).to(beFalse())
+                    expect(view.carouselView.carouselPageControl.isHidden).to(beFalse())
+                }
+
                 it("it will not display Carousel view and Carousel Page Control if the data is valid but campaign has header and body text") {
 
                     view.setup(viewModel: TestHelpers.generateFullViewModel(carouselData: carouselData))
@@ -339,6 +346,13 @@ class BaseViewPresenterMock: BaseViewPresenterType {
 }
 
 class FullViewPresenterMock: FullViewPresenterType {
+    var isValidSize = false
+    var isValidPosition = false
+
+    func validateAndAdjustModifyModal(modal: ResizeableModal?) -> (isValidSize: Bool, isValidPosition: Bool, updatedModel: ResizeableModal?) {
+        return (isValidSize, isValidPosition, modal)
+    }
+
     var carouselData: [CarouselData]?
     var view: FullViewType?
     var campaign: Campaign!